bot_saves_princess.py3
Scenario:
Princess Peach is trapped in one of the four corners of a square grid. You are in the centre of the grid and can move one step at a time in any of the four directions. 

Input format:
The first line contains an odd integer N (3 <= N < 100) denoting the size of the grid. This is followed by an NxN grid. Each cell is denoted by '-' (ascii value: 45). The bot position is denoted by 'm' and the princess position is denoted by 'p'.

Grid is indexed using Matrix Convention

Output format:
Print out the moves you will take to rescue the princess in one go. The moves must be separated by '\n', a newline. The valid moves are LEFT or RIGHT or UP or DOWN.